requirements.txt
fastapi<1
uvicorn<1
slack_sdk>=3.25.0,<4
slack_bolt>=1.19.0
pydantic-settings>=2.3.4
aiohttp>=3.8.6
certifi==2023.7.22
filelock>=3.15.4
schedule>=1.2.2